pyproject.toml
[tool.poetry]
name = "shredderpy"
version = "0.1.0"
description = "Shred images and paste them back together... kind of."
license = "MIT"
authors = ["Jason Hutson <110316760+HutsonJason@users.noreply.github.com>"]
repository = "https://github.com/HutsonJason/shredderpy"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.4.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"
mkdocs-material = "^9.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
